requirements.txt
soda-core==3.1.0
soda-core-contracts==3.1.0
soda-core-vertica==3.1.0
pyyaml